requirements.txt
asgiref==3.7.2
cffi==1.15.1
Django==4.2.4
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
